+++ LegendWeathers/Config.cs
@@ -15,9 +15,12 @@ namespace LegendWeathers
         public readonly ConfigEntry<string> majoraMaskValue;
         public (int, int) majoraMaskValueParsed;
 
+        public readonly ConfigEntry<bool> bloodMoonWeather;
+
         public Config(ConfigFile cfg)
         {
             cfg.SaveOnConfigSet = false;
+
             majoraWeather = cfg.Bind("Majora Moon", "Enabled", true, "Enable the Majora Moon weather.");
             majoraMoonModel = cfg.Bind("Majora Moon", "Model version", "3DS", new ConfigDescription("Choose the model version of the moon, if you want a more retro look try the N64 version.\nOther models are also available for fun !", new AcceptableValueList<string>("3DS", "N64", "Faceless", "Boomy", "Owl", "Abibabou", "Joy", "Dice", "Baldy")));
             majoraMoonModelAutomatic = cfg.Bind("Majora Moon", "Automatic model selection", false, "Allows the model to be automatically adjusted on moons based on certain conditions, this will vary depending on your installed mods.");
@@ -25,6 +28,9 @@ namespace LegendWeathers
             majoraOcarinaCompatible = cfg.Bind("Majora Moon", "Ocarina compatibility", true, "If you have the Ocarina item, playing Oath to Order while in altitude when the moon is about to crash will start a special animation.\nWill not work if ChillaxScraps is not installed.");
             majoraCompanyCompatible = cfg.Bind("Majora Moon", "Company Moon compatibility", false, "By default, the Majora Moon can't spawn on Company Moons but activating this config will make this a reality.\nYou may also need to remove the company moons in the MajoraMoon blacklist section in the WeatherRegistery config file.");
             majoraMaskValue = cfg.Bind("Majora Moon", "Majora Mask Item value", "200,400", "The min,max scrap value of the Majora's Mask item, supposed to be very high.\nThe final value will be randomized between these 2 numbers, but not divided by any external factors.");
+
+            bloodMoonWeather = cfg.Bind("Blood Moon", "Enabled", true, "Enable the Blood Moon weather.");
+
             cfg.Save();
             cfg.SaveOnConfigSet = true;
         }